python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
openai==1.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
torch==2.1.0
numpy==1.24.3
anthropic>=0.30
